refactor(walletd): address Gemini review feedback on API key auth
- Use i64 for api_key.id throughout (schema BigInt, model, SDK types,
  handler signatures) — SQLite INTEGER PRIMARY KEY is 64-bit
- api_keys_insert: fetch by key_hash instead of ORDER BY id DESC to
  avoid race in concurrent inserts
- Extract api_keys_get_by_hash_conn / api_keys_list_all_conn helpers
  to eliminate duplicated impl between WriteTransaction and ReadTransaction
- Login handler: match on AuthCredentials to move raw key string directly
  into Zeroizing<String> — avoids transient non-zeroized copy
- Add now_unix() helper in handlers.rs; replace duplicate inline
  SystemTime::now() expressions in login and revoke handlers
- Explicit drop(encoded) before returning AuthCreateApiKeyResponse
  to zeroize the raw key before the response escapes the function

+++ clients/wallet_daemon_client/src/types.rs
@@ -1484,7 +1484,7 @@ pub struct AuthCreateApiKeyRequest {
 #[derive(Debug, Clone, Deserialize, Serialize)]
 #[cfg_attr(feature = "ts", derive(ts_rs::TS), ts(export, export_to = "wallet-types/"))]
 pub struct AuthCreateApiKeyResponse {
-    pub id: i32,
+    pub id: i64,
     pub name: String,
     /// The raw API key — only returned once. Store securely.
     pub key: String,
@@ -1495,7 +1495,7 @@ pub struct AuthCreateApiKeyResponse {
 #[derive(Debug, Clone, Deserialize, Serialize)]
 #[cfg_attr(feature = "ts", derive(ts_rs::TS), ts(export, export_to = "wallet-types/"))]
 pub struct ApiKeyInfo {
-    pub id: i32,
+    pub id: i64,
     pub name: String,
     pub permissions: Vec<String>,
     pub created_at: i64,
@@ -1516,7 +1516,7 @@ pub struct AuthListApiKeysResponse {
 #[derive(Debug, Clone, Deserialize, Serialize)]
 #[cfg_attr(feature = "ts", derive(ts_rs::TS), ts(export, export_to = "wallet-types/"))]
 pub struct AuthRevokeApiKeyRequest {
-    pub id: i32,
+    pub id: i64,
 }
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
